Isolate execute_agent_functor in its own header
