+++ agency/detail/bulk_functions/bulk_async.hpp
@@ -4,6 +4,7 @@
 #include <agency/detail/type_traits.hpp>
 #include <agency/detail/bulk_functions/decay_parameter.hpp>
 #include <agency/detail/bulk_functions/unpack_shared_parameters_from_executor_and_invoke.hpp>
+#include <agency/detail/bulk_functions/execute_agent_functor.hpp>
 #include <utility>
 
 namespace agency
